Move home component into home folder
