+++ client/src/App.js
@@ -1,7 +1,7 @@
 
 import './App.css';
 import { Routes, Route } from "react-router-dom";
-import Home from './pages/Home';
+import Home from './pages/home/Home';
 
 function App() {
   return (
